correction of error in the communication with the Database

+++ src/main/java/br/com/kleryton/api/services/UserService.java
@@ -4,6 +4,6 @@ import br.com.kleryton.api.domain.User;
 
 public interface UserService {
 
-    User findByID(Integer id);
+    User findByID(Long id);
 
 }
